checkers: extend equalFold checker to bytes.EqualFold
Handle bytes comparisons in addition to strings comparisons.

Also fix a bug that made linter suggest invalid replacements.
It happened because we haven't checked which call is under
callX and callY. If there is something that is not ToUpper
or ToLower, we should not simplify a call to its first argument.

In addition, add negative tests.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>