add regexpPattern checker
Currently only checks whether the pattern contains
an unescaped "." near known top level domain.

So, `google.com` is probably flawed and
`google\.com` is considered to be correct.

It's not a security issue most of the time, but
it's better to keep that surface as thin as possible
to avoid potential problems.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>