checkers: fix integration tests
They were failing due to the lack of unlambda warning.
It happened because the code used a closure (var-assigned func)
instead of a normal func. We don't warn on closures
anymore (see #888), so that warning disappeared.

For whatever reason, CI passed and we didn't notice it.
Maybe integration tests are not executed as a part of our pipeline?

Anyway, tests should be fixed now.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>