checkers: add ruleguard checker
This is part of the ruleguard integration experiment.

Hopefully, we can get some feedback to improve the way it's integrated.

The simplest way to try it out:
	gocritic check -enable ruleguard -@ruleguard.rules <gorules> <target>

Where:
	<gorules> is a path to a Go rules file
	<target> is a linting target

An example gorules file can be found here:
	https://github.com/quasilyte/go-ruleguard/blob/master/rules.go

Visit https://github.com/quasilyte/go-ruleguard for more info on the ruleguard.

Updates #873
Updates #869
Updates #837
Updates #836
Updates #831

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>