all: integrate lintpack as internal framework
Instead of using a lintpack as a dependency, clone
most of its code that we need.

This is an intermediate step to make transition to
go/analysis easier. Now we don't need to keep
lintpack dependency in sync and can make
framework/linter compatible with go/analysis later.

Doing a v0.5.0 release is suggested.

Updates #854

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>