checkers: make it possible to write embedded ruleguard rules
This change makes several steps to make it possible to use
ruleguard rules to express new go-critic checkers.

* `checkers/rules` contains a ruleguard rules; these rules will
  be converted into go-critic checkers on-the-fly

* `checkers/rulesdata` is auto-generated fron the `checkers/rules`
  with a help of `go-bindata` (`go:embed` requires a Go 1.16 that is too recent)

* `checkers/checkers.go` reads `checkers/rulesdata` to parse ruleguard rules;
  then it turns them into ordinary go-critic checkers

These auto-generated checkers follow all rules of normal checkers.
They're tested in the same way and can be disabled separately.

To give an example of how it could be used, I've included
a simple rule `httpNoBody` proposed by @cristaloleg.

The only peculiar part about this implementation is that we
create a dedicated `ruleguard.Engine` for every checker.
It's as good as it gets for now.

The reason behind this is go-critic public API for checkers.
We're expected to run a checker over a file and give its reports back.
ruleguard optimized to work differently: all checkers are merged as
one executable unit.