README.md: Doc sec tag disabled, fails if enabled
Document that the security tag is disabled by default. Also document
that there are currently no checks in the security tag, so enabling it
will fail.

$ ./gocritic check -v -enable='#security
init checkers: empty checkers set selected
$ echo $?
1

When trying to enable the tag through golangci-lint, this is the error:

ERRO Invalid gocritic settings: gocritic [enabled]tag "security" doesn't
exist

The tag was added in #989.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>