checkers: remove experimental tag from ioutil checker
The ioutil package was deprecated as part of go 1.16.

Now that go 1.18 has dropped, the core go team no longer supports
1.16.

So it should be safe to remove the experimental tag. This will switch
the check from disabled by default to enabled by default.

Note that the check still has internal if statements to only fire on code
running 1.16 or newer, so even though it's enabled by default, it won't
start complaining on code running older versions of go.